Align weather data persistence and refactor seeding to idempotent UPSERT. Exclude documentation files.

+++ app/database/queries/weather_queries.py
@@ -49,8 +49,8 @@ UPSERT_WEATHER_HOURLY = """
         breach_metric,
         breach_observed_value
     ) VALUES (
-        $1, $2, $3, $4, $5, $6, $7,
-        ST_SetSRID(ST_MakePoint($7, $6), 4326),
+        $1, $2, $3, $4, $5, $6::numeric, $7::numeric,
+        ST_SetSRID(ST_MakePoint($7::float8, $6::float8), 4326),
         $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, 
         $18, $19, $20, $21, $22, $23, $24, $25, $26,
         $27, $28, $29, $30, $31, $32, $33, $34, $35, 
@@ -124,8 +124,8 @@ UPSERT_WEATHER_DAILY = """
         flag_cold_wave_day,
         worst_breach_severity
     ) VALUES (
-        $1, $2, $3, $4, $5, $6, $7,
-        ST_SetSRID(ST_MakePoint($7, $6), 4326),
+        $1, $2, $3, $4, $5, $6::numeric, $7::numeric,
+        ST_SetSRID(ST_MakePoint($7::float8, $6::float8), 4326),
         $8, $9, $10, $11, $12, $13, $14, $15, $16, 
         $17, $18, $19, $20, $21, $22, $23, $24, $25, $26, $27
     )
